rename core.newapplication -> core.application
