+++ IPython/parallel/apps/iploggerapp.py
@@ -25,7 +25,7 @@ import sys
 
 import zmq
 
-from IPython.core.newapplication import ProfileDir
+from IPython.core.application import ProfileDir
 from IPython.utils.traitlets import Bool, Dict, Unicode
 
 from IPython.parallel.apps.baseapp import (
